# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FlowReport API
  description: |
    FlowReport - ReportingOps SaaS 플랫폼 API 문서

    이 문서는 FlowReport의 주요 API 엔드포인트를 정의합니다.

    ## 인증
    모든 API는 세션 기반 인증이 필요합니다. NextAuth.js를 통해 인증됩니다.

    ## 권한
    - VIEWER: 조회 권한
    - EDITOR: 조회 및 수정 권한
    - ADMIN: 모든 권한 (생성, 수정, 삭제 포함)
  version: 1.0.0
  contact:
    name: FlowReport Team
servers:
  - url: /api
    description: API 기본 경로
...
paths:
...
  /workspaces/{workspaceId}/action-templates:
    get:
      summary: 액션 템플릿 목록 조회
      description: |
        워크스페이스의 활성화된 액션 템플릿 목록을 조회합니다.
        triggerType과 department로 필터링할 수 있습니다.
      operationId: listActionTemplates
      tags:
        - ActionTemplates
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: triggerType
          in: query
          description: 트리거 타입으로 필터링
          required: false
          schema:
            $ref: '#/components/schemas/TriggerType'
        - name: department
          in: query
          description: 담당 부서로 필터링
          required: false
          schema:
            $ref: '#/components/schemas/Department'
      responses:
        '200':
          description: 액션 템플릿 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActionTemplate'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: 액션 템플릿 생성
      description: |
        새로운 액션 템플릿을 생성합니다.
        동일한 triggerType과 triggerLevel 조합은 워크스페이스 내에서 고유해야 합니다.
      operationId: createActionTemplate
      tags:
        - ActionTemplates
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActionTemplateRequest'
      responses:
        '201':
          description: 액션 템플릿 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    $ref: '#/components/schemas/ActionTemplate'
        '400':
          description: 필수 필드 누락
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Missing required fields: triggerType, title, description, department, departmentUrl"
        '409':
          description: 중복된 템플릿 존재
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Template with this triggerType and triggerLevel already exists"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workspaces/{workspaceId}/action-templates/{templateId}:
    get:
      summary: 액션 템플릿 상세 조회
      description: 특정 액션 템플릿의 상세 정보를 조회합니다.
      operationId: getActionTemplate
      tags:
        - ActionTemplates
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: 액션 템플릿 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    $ref: '#/components/schemas/ActionTemplate'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: 액션 템플릿 수정
      description: |
        기존 액션 템플릿을 수정합니다.
        기본 템플릿(isDefault=true)의 경우 triggerType, triggerLevel, department 필드는 수정할 수 없습니다.
      operationId: updateActionTemplate
      tags:
        - ActionTemplates
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActionTemplateRequest'
      responses:
        '200':
          description: 액션 템플릿 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    $ref: '#/components/schemas/ActionTemplate'
        '403':
          description: 기본 템플릿의 보호된 필드 수정 시도
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Cannot modify protected fields of default templates"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: 액션 템플릿 삭제
      description: |
        액션 템플릿을 삭제합니다.
        기본 템플릿(isDefault=true)은 삭제할 수 없으며, 대신 비활성화(isActive=false)해야 합니다.
      operationId: deleteActionTemplate
      tags:
        - ActionTemplates
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: 액션 템플릿 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '403':
          description: 기본 템플릿 삭제 시도
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Cannot delete default templates. Deactivate instead."
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ===========================================
  # Competitor API
  # ===========================================
...
components:
...
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      description: 워크스페이스 ID
      required: true
      schema:
        type: string
        example: "clxxxxxxxxxxxxxxxxxx"
    TemplateId:
      name: templateId
      in: path
      description: 액션 템플릿 ID
      required: true
      schema:
        type: string
        example: "clxxxxxxxxxxxxxxxxxx"
...
  responses:
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Template not found"
...
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"

  # ===========================================
  # Schemas
  # ===========================================
  schemas:
...
    TriggerType:
      type: string
      description: 액션 트리거 타입
      enum:
        - REVENUE_DECLINE_CRITICAL
        - REVENUE_DECLINE_WARNING
        - ENGAGEMENT_DECLINE_CRITICAL
        - ENGAGEMENT_DECLINE_WARNING
        - CONVERSION_LOW_CRITICAL
        - CONVERSION_LOW_WARNING
        - CHANNEL_METRIC_DECLINE
    AlertLevel:
      type: string
      description: 알림 레벨
      enum:
        - CRITICAL
        - WARNING
        - INFO
    ActionPriority:
      type: string
      description: 액션 우선순위
      enum:
        - HIGH
        - MEDIUM
        - LOW
    Department:
      type: string
      description: 담당 부서
      enum:
        - MARKETING
        - COMMERCE
        - OVERALL
...
    Error:
      type: object
      description: 에러 응답
      properties:
        error:
          type: string
          description: 에러 메시지
      required:
        - error
...
    ActionTemplate:
      type: object
      description: 액션 템플릿
      properties:
        id:
          type: string
          description: 템플릿 고유 ID
          example: "clxxxxxxxxxxxxxxxxxx"
        workspaceId:
          type: string
          description: 워크스페이스 ID
        triggerType:
          $ref: '#/components/schemas/TriggerType'
        triggerLevel:
          $ref: '#/components/schemas/AlertLevel'
        title:
          type: string
          description: 템플릿 제목
          example: "매출 급감 대응 체크리스트"
        description:
          type: string
          description: 템플릿 설명
          example: "매출이 전주 대비 20% 이상 감소했을 때 확인해야 할 사항"
        priority:
          $ref: '#/components/schemas/ActionPriority'
        department:
          $ref: '#/components/schemas/Department'
        departmentUrl:
          type: string
          description: 담당 부서 관련 URL
          example: "https://slack.com/channel/commerce"
        steps:
          type: array
          description: 액션 단계 목록
          items:
            type: string
          example:
            - "1. 쿠폰/프로모션 상태 확인"
            - "2. 상품 재고 확인"
            - "3. 경쟁사 프로모션 확인"
        isActive:
          type: boolean
          description: 활성화 여부
          example: true
        isDefault:
          type: boolean
          description: 시스템 기본 템플릿 여부
          example: false
        createdAt:
          type: string
          format: date-time
          description: 생성일시
        updatedAt:
          type: string
          format: date-time
          description: 수정일시
    CreateActionTemplateRequest:
      type: object
      description: 액션 템플릿 생성 요청
      required:
        - triggerType
        - title
        - description
        - department
        - departmentUrl
      properties:
        triggerType:
          $ref: '#/components/schemas/TriggerType'
        triggerLevel:
          $ref: '#/components/schemas/AlertLevel'
          default: WARNING
        title:
          type: string
          description: 템플릿 제목
          minLength: 1
        description:
          type: string
          description: 템플릿 설명
          minLength: 1
        priority:
          $ref: '#/components/schemas/ActionPriority'
          default: MEDIUM
        department:
          $ref: '#/components/schemas/Department'
        departmentUrl:
          type: string
          description: 담당 부서 관련 URL
        steps:
          type: array
          description: 액션 단계 목록
          items:
            type: string
          default: []
    UpdateActionTemplateRequest:
      type: object
      description: 액션 템플릿 수정 요청
      properties:
        title:
          type: string
          description: 템플릿 제목
        description:
          type: string
          description: 템플릿 설명
        priority:
          $ref: '#/components/schemas/ActionPriority'
        departmentUrl:
          type: string
          description: 담당 부서 관련 URL
        steps:
          type: array
          description: 액션 단계 목록
          items:
            type: string
        isActive:
          type: boolean
          description: 활성화 여부

    # --- Competitor Schemas ---
...